requirements.txt
cryptography==38.0.1
gspread==5.4.0
oauth2client==4.1.3
Pillow==9.2.0
PyAutoGUI==0.9.53
PyGithub==1.56
pystray==0.19.4
requests==2.28.1
